Update README.md
Write a short program that computes the volume of a sphere given the radius.

Your program should have a prompt asking the user to input the value of the radius, calculate the the volume of the sphere, and then output the volume to the console with a label.

Output should be rounded to 2 decimal places. Use Wikipedia to find the equation for the sphere volume. Use a named constants for PI instead of hard coding 3.14 in your math expression. Also try to use good variable names.

Using a calculator work out the solution for a certain radius. Verify that you get the correct result with your program. Don't forget that dividing two integers (like 4/3) does not produce a result with a decimal (like 1.33). To do floating-point division the numbers need decimal points (like 4.0/3.0).